+++ packages/rpc_dart_transports/lib/src/transports/_index.dart
@@ -2,6 +2,7 @@
 //
 // SPDX-License-Identifier: MIT
 
+export 'http/_index.dart';
 export 'http2/_index.dart';
 export 'isolate/isolate_transport.dart';
 export 'websocket/_index.dart';