+++ packages/rpc_dart_transports/lib/src/transports/http/_index.dart
@@ -1,0 +1,7 @@
+// SPDX-FileCopyrightText: 2025 Karim "nogipx" Mamatkazin <nogipx@gmail.com>
+//
+// SPDX-License-Identifier: MIT
+
+export 'rpc_http1_caller_transport.dart';
+export 'rpc_http1_responder_transport.dart';
+export 'rpc_http1_server.dart';